LayoutTests/platform/gtk/imported/w3c/web-platform-tests/pointerevents/persistentDeviceId/get-persistendeviceid-from-pointer-mouse-event.tentative-expected.txt


Harness Error (FAIL), message = Error: assert_equals: click deviceId is 0 expected (number) 0 but got (undefined) undefined

PASS PointerEvent.persistentDeviceId
